10. Regular Expression Matching
Given an input string s and a pattern p, implement regular expression matching with support for '.' and '*' where:

'.' Matches any single character.
'*' Matches zero or more of the preceding element.
The matching should cover the entire input string (not partial).

if is x*, then we can give up this element so
dp[i][j] can be dp[i-1][j]
and can match any nunmber of x
so if s[j-1] == x: dp[i][j] can be dp[i][j-1]
note that by "can be" it means the results are ORed together