992. Subarrays with K Different Integers
Given an array nums of positive integers, call a
(contiguous, not necessarily distinct) subarray of nums
good if the number of different integers in that
subarray is exactly k.

(For example, [1,2,3,1,2] has 3 different integers: 1, 2, and 3.)

Return the number of good subarrays of nums.

we keep a counter "repeat" to check if there are multiple possibilities of left
and add repeat + 1 for each right
clear the repeat if the set has changed